// src/domain/settings/settings-schema.ts
export type StoreSettings = {
  // General Info
  siteName: string;
  siteDescription: string;
  siteContactEmail: string;
  sitePhone: string;
  siteHours: string;
  whatsappUrl: string;
  messengerUrl: string;

  // Delivery & Location
  locationName: string;
  deliveryDhaka: string;
  deliveryOutside: string;
  deliveryNote: string;
  courierPartners: string[];

  // Social Links
  fbPageUrl: string;
  fbUrl: string;
  igUrl: string;
  ytUrl: string;
};